Module siphasher::sip128

An implementation of SipHash with a 128-bit output.

Macros

compress 🔒

load_int_le 🔒

Loads an integer of the desired type from a byte stream, in LE order. Uses copy_nonoverlapping to let the compiler generate the most efficient way to load it from a possibly unaligned address.

Structs

Hash128

A 128-bit (2x64) hash output

Hasher 🔒

Sip13Rounds 🔒

Sip24Rounds 🔒

SipHasher

An implementation of SipHash128 2-4.

SipHasher13

An implementation of SipHash128 1-3.

SipHasher24

An implementation of SipHash128 2-4.

State 🔒

Traits

Hasher128

Functions

u8to64_le 🔒 ⚠

Loads a u64 using up to 7 bytes of a byte slice. It looks clumsy but the copy_nonoverlapping calls that occur (via load_int_le!) all have fixed sizes and avoid calling memcpy, which is good for speed.