Struct siphasher::sip128::SipHasher

pub struct SipHasher(_);

An implementation of SipHash128 2-4.

SipHash is a general-purpose hashing function: it runs at a good speed (competitive with Spooky and City) and permits strong keyed hashing. This lets you key your hashtables from a strong RNG, such as rand::os::OsRng.

Although the SipHash algorithm is considered to be generally strong, it is not intended for cryptographic purposes. As such, all cryptographic uses of this implementation are strongly discouraged.

Methods

impl SipHasher

fn new() -> SipHasher

Creates a new SipHasher with the two initial keys set to 0.

fn new_with_keys(key0: u64, key1: u64) -> SipHasher

Creates a SipHasher that is keyed off the provided keys.

fn keys(&self) -> (u64, u64)

Get the keys used by this hasher