Struct siphasher::sip128::SipHasher13

pub struct SipHasher13 { /* private fields */ }

An implementation of SipHash128 1-3.

Implementations

impl SipHasher13

pub fn new() -> SipHasher13

Creates a new SipHasher13 with the two initial keys set to 0.

pub fn new_with_keys(key0: u64, key1: u64) -> SipHasher13

Creates a SipHasher13 that is keyed off the provided keys.

pub fn new_with_key(key: &[u8; 16]) -> SipHasher13

Creates a SipHasher13 from a 16 byte key.

pub fn keys(&self) -> (u64, u64)

Get the keys used by this hasher

pub fn key(&self) -> [u8; 16]

Get the key used by this hasher as a 16 byte vector

pub fn hash(&self, bytes: &[u8]) -> Hash128

Hash a byte array - This is the easiest and safest way to use SipHash.